added def for merge functions some cleaning

+++ feature_extractor.py
@@ -23,9 +23,9 @@ dst = "2.wav"
 mp3_dir = "dataset//DEAM_audio//MEMD_audio"
 wav_dir = "dataset//DEAM_audio//wav_audio"
 csv_dir = "dataset//extracted_features//sampling"
-csv_file = csv_dir + "//features_no_sampling.csv"
+csv_file = "dataset//extracted_features//features_no_sampling.csv"
 #directory where dynamic annotations per frame are stored
-dyn_ann_dir = "dataset//DEAM_annotations//annotations//dynamic//annotations averaged per song//dynamic (per second annotations)//"
+dyn_ann_dir = "dataset//DEAM_annotations//annotations//dynamic//"
 #directory where static annotations per song are stored
 stat_ann_dir = "dataset//DEAM_annotations//annotations//song_level//"
 
@@ -51,31 +51,28 @@ def convert_mp3_to_wav(mp3_dir, wav_dir):
     for the number of specified frames
     Converts stereo to mono signal
     Saves extracted features into csv files in csv_dir.
-    One csv files per song. Columns represent the features, rows - the frames.
+    One csv files per song, title is the song_id. Columns represent the features, rows - the frames.
     The first row is a header row with feature names
     The first column is the frame counter """
 def extract_features_per_frame(wav_dir, csv_dir, nr_frames):              
-    song_size = 45
-    window_size = 500
-
-    #perform audio feature extraction using ShortTermFeatures:
-    #output: a csv file for each song where the title is the song ID, columns correspond to features and rows to windows
+    #song_size = 45000
+    #window_size = 500  
     with os.scandir(wav_dir) as dir:
         for song in dir:
-            if song.is_file() and song.name.endswith(".wav"):
+            song_id = song.name.rstrip(".wav")
+            if song.is_file() and song.name.endswith(".wav") and not os.path.exists(csv_dir + '/' + song_id + '.csv'):
                 print("Processing song: " + song.name)           
                 sample_rate, signal = audioBasicIO.read_audio_file(song)
                 print('Sample rate: ', sample_rate)
                 print("Number of samples: ",len(signal))
+                #perform conversion from stereo to mono
                 signal = audioBasicIO.stereo_to_mono(signal)
-                #sample rate retrieved from read audio file is different that sample rate used as a second arg in feature extraction
-                features_and_deltas, feature_names = ShortTermFeatures.feature_extraction(signal, sample_rate, len(signal)/nr_frames, len(signal)/nr_frames) #TODO: make this work
-                #print(len(features_and_deltas))
-                #print(features_and_deltas)
-                features = np.transpose(features_and_deltas[:34,:]) # why is it limited to 34
-                #print(features)
-                song_id = song.name.rstrip(".wav")
+                #perform audio feature extraction using ShortTermFeatures:
+                features_and_deltas, feature_names = ShortTermFeatures.feature_extraction(signal, sample_rate, len(signal)/nr_frames, len(signal)/nr_frames) 
+                features = np.transpose(features_and_deltas[:34,:])
+                #output: a csv file for each song where the title is the song ID, columns correspond to features and rows to windows
                 with open(file=csv_dir + '/' + song_id + '.csv', mode='w', encoding='ASCII') as file:
+                    print("Writing " + csv_dir + '/' + song_id + '.csv')
                     file.write('frame #' + np_row_to_string(feature_names[:34]) + '\n')
                     current_frame = 0
                     for row in features:
@@ -116,15 +113,24 @@ def extract_features_per_song(wav_dir, csv_file):
                     file.write(song_id + np_row_to_string(features) + '\n')
 
 
+
+
 """ Converts a row from a numpy array to a string with the format
     ,element_1,element_2,...,element_n
-    Used when printing to files"""
+    Used when printing to files""" 
 def np_row_to_string(row):
     result = ''
     for element in row:
         result += ',' + str(element)
 
     return result
+
+""" Performs merging csv files: csvs with features extracted per frame and dynamic annotations with arousal/valence per frame
+The resulting csv file has format:
+
+songId_frameID  Feature1    Feature2    ... FeatureN    Arousal Valence"""
+
+#def merge_features_annontations_per_frame(features_dir, stat_ann_dir):
 
 """ Performs merging csv files: features extracted per song and static annotations with arousal/valence per song
 The resulting csv file has format:
@@ -134,10 +140,8 @@ songID  Feature1    Feature2    ... FeatureN    Arousal Valence"""
 #def merge_features_annontations_per_song(features_dir, stat_ann_dir):
 
 #convert_mp3_to_wav(mp3_dir, wav_dir)
-#extract_features_per_frame(wav_dir, csv_dir, 90)
-extract_features_per_song(wav_dir, csv_file)
-#extract_features_per_song(wav_dir, csv_file)
 extract_features_per_frame(wav_dir, csv_dir, 90)
+#extract_features_per_song(wav_dir, csv_file)
 
 #there should be two variants of feature extraction (in two separate scripts): one that uses frame size of 500 msec (sampling rate of 2 Hz) and feature extraction per song level - using entire song (45 sec) as a frame size
